+++ src/index.ts
@@ -1,9 +1,6 @@
 export * from './Die'
 export { default as roll } from './roll'
 export { default as validateDiceNotation } from './validateDiceNotation'
-export {
-  parseRollArgument,
-  default as parseRollArguments
-} from './parseRollArguments'
+export { default as parameterizeRollArgument } from './parameterizeRollArgument'
 export { default as generateRollResult } from './roll/generateRollResult'
 export * from './types'
