+++ src/roll/index.ts
@@ -1,6 +1,6 @@
 import { RollArgument, RollResult } from '~types'
 import generateRollResult from './generateRollResult'
-import parseRollArguments from '~src/parseRollArguments'
+import parseRollArguments from './parseRollArguments'
 
 export default function roll(arg?: RollArgument): RollResult {
   return generateRollResult(parseRollArguments(arg))
