+++ src/utils/formatDescription/stringFormatters.ts
