+++ src/utils/parseModifiers.ts
