+++ src/utils/parseNotation.ts
